+++ TALLER_FELIPEVILLAMIL_JULIANMURCIA/ParameterV3/src/co/edu/uptc/model/Product.java
@@ -2,10 +2,10 @@ package co.edu.uptc.model;
 
 public class Product {
     private String description;
-    private double price;
+    private int price;
     private String unit;
 
-    public Product(String description, double price, String unit) {
+    public Product(String description, int price, String unit) {
         this.description = description;
         this.price = price;
         this.unit = unit;
@@ -25,6 +25,6 @@ public class Product {
 
     @Override
     public String toString() {
-        return String.format("%-20s | $%-10.2f | %s", description, price, unit);
+        return String.format("%-20s | $%-10.2f | %s", description, (double)price, unit);
     }
 }
